requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
pytz==2018.5
requests==2.19.1
urllib3==1.23
